Yandex.Realestate analysis
Data used: the data of the Yandex.Realty service - an archive of ads for the sale of apartments in St. Petersburg and neighbouring settlements for several years.
Task: We need to learn how to determine the market value of real estate. The job is to set the parameters. This will allow to build an automated system: it will track anomalies and fraudulent activity.