Game sales analysis
Strimchik online store, which sells computer games all over the world, wants to identify the patterns that determine the success of the game. Historical game sales data, user and expert ratings, genres and platforms (such as Xbox or PlayStation) are available from public sources. This will allow to bid on a potentially popular product and plan advertising campaigns.